+++ day03/ex01/FragTrap.cpp
@@ -95,7 +95,7 @@ void	FragTrap::takeDamage(unsigned int amount)
 
 void	FragTrap::beRepaired(unsigned int amount)
 {
-	if (this->_hitPoints + amount >= this->_maxHitPoints)
+	if (this->_hitPoints + (int)amount >= this->_maxHitPoints)
 	{
 		this->_hitPoints = this->_maxHitPoints;
 		std::cout << this->_name << " is fully repaired!" << std::endl;
